% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{tous_les_programmes}
\alias{tous_les_programmes}
\title{liste tous les Rmd candidats au deploiement contenus dans le package}
\usage{
tous_les_programmes(pattern = "*.Rmd$")
}
\arguments{
\item{pattern}{pattern des fichiers a proposer}
}
\description{
liste tous les Rmd candidats au deploiement contenus dans le package
}
